+++ TwitchDownloaderWPF/PageChatDownload.xaml.cs
@@ -8,11 +8,12 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media.Imaging;
-using TwitchDownloaderWPF.Properties;
 using TwitchDownloaderCore;
 using TwitchDownloaderCore.Chat;
 using TwitchDownloaderCore.Options;
 using TwitchDownloaderCore.TwitchObjects.Gql;
+using TwitchDownloaderWPF.Properties;
+using TwitchDownloaderWPF.Services;
 using WpfAnimatedGif;
 
 namespace TwitchDownloaderWPF
@@ -108,12 +109,12 @@ namespace TwitchDownloaderWPF
                         try
                         {
                             string thumbUrl = taskVideoInfo.Result.data.video.thumbnailURLs.FirstOrDefault();
-                            imgThumbnail.Source = await InfoHelper.GetThumb(thumbUrl);
+                            imgThumbnail.Source = await ThumbnailService.GetThumb(thumbUrl);
                         }
                         catch
                         {
                             AppendLog(Translations.Strings.ErrorLog + Translations.Strings.UnableToFindThumbnail);
-                            var (success, image) = await InfoHelper.TryGetThumb(InfoHelper.THUMBNAIL_MISSING_URL);
+                            var (success, image) = await ThumbnailService.TryGetThumb(ThumbnailService.THUMBNAIL_MISSING_URL);
                             if (success)
                             {
                                 imgThumbnail.Source = image;
@@ -158,12 +159,12 @@ namespace TwitchDownloaderWPF
                         try
                         {
                             string thumbUrl = taskClipInfo.Result.data.clip.thumbnailURL;
-                            imgThumbnail.Source = await InfoHelper.GetThumb(thumbUrl);
+                            imgThumbnail.Source = await ThumbnailService.GetThumb(thumbUrl);
                         }
                         catch
                         {
                             AppendLog(Translations.Strings.ErrorLog + Translations.Strings.UnableToFindThumbnail);
-                            var (success, image) = await InfoHelper.TryGetThumb(InfoHelper.THUMBNAIL_MISSING_URL);
+                            var (success, image) = await ThumbnailService.TryGetThumb(ThumbnailService.THUMBNAIL_MISSING_URL);
                             if (success)
                             {
                                 imgThumbnail.Source = image;
