+++ TwitchDownloaderWPF/PageVodDownload.xaml.cs
@@ -17,6 +17,7 @@ using TwitchDownloaderCore;
 using TwitchDownloaderCore.Options;
 using TwitchDownloaderCore.TwitchObjects.Gql;
 using TwitchDownloaderWPF.Properties;
+using TwitchDownloaderWPF.Services;
 using WpfAnimatedGif;
 
 namespace TwitchDownloaderWPF
@@ -80,12 +81,12 @@ namespace TwitchDownloaderWPF
                     try
                     {
                         string thumbUrl = taskVideoInfo.Result.data.video.thumbnailURLs.FirstOrDefault();
-                        imgThumbnail.Source = await InfoHelper.GetThumb(thumbUrl);
+                        imgThumbnail.Source = await ThumbnailService.GetThumb(thumbUrl);
                     }
                     catch
                     {
                         AppendLog(Translations.Strings.ErrorLog + Translations.Strings.UnableToFindThumbnail);
-                        var (success, image) = await InfoHelper.TryGetThumb(InfoHelper.THUMBNAIL_MISSING_URL);
+                        var (success, image) = await ThumbnailService.TryGetThumb(ThumbnailService.THUMBNAIL_MISSING_URL);
                         if (success)
                         {
                             imgThumbnail.Source = image;
