+++ TwitchDownloaderWPF/Services/ThumbnailService.cs
@@ -2,9 +2,9 @@
 using System.Threading.Tasks;
 using System.Windows.Media.Imaging;
 
-namespace TwitchDownloaderWPF
+namespace TwitchDownloaderWPF.Services
 {
-    class InfoHelper
+    public static class ThumbnailService
     {
         public const string THUMBNAIL_MISSING_URL = @"https://vod-secure.twitch.tv/_404/404_processing_320x180.png";
         private static readonly HttpClient _httpClient = new();
